work/generated/blk-motor/implementation.py
# Requirement: RQ-MD-001
# Description: Elevator motor control system for vertical movement

import logging
import time
import threading
from typing import Any

logging.basicConfig(level=logging.INFO)


class MotorControl:
    """
    Class to control the motor of an elevator.
    """

    def __init__(self) -> None:
        self.speed = 0
        self.temperature = 0
        self.current = 0
        self.direction = 0
        self.position = 0
        self.stop = False
        self.lock = threading.RLock()

    def set_speed(self, speed: int) -> None:
        """
        Set the speed of the motor.
        """
        with self.lock:
            self.speed = speed

    def get_speed(self) -> int:
        """
        Get the speed of the motor.
        """
        with self.lock:
            return self.speed

    def set_temperature(self, temperature: int) -> None:
        """
        Set the temperature of the motor.
        """
        with self.lock:
            self.temperature = temperature

    def get_temperature(self) -> int:
        """
        Get the temperature of the motor.
        """
        with self.lock:
            return self.temperature

    def set_current(self, current: int) -> None:
        """
        Set the current of the motor.
        """
        with self.lock:
            self.current = current

    def get_current(self) -> int:
        """
        Get the current of the motor.
        """
        with self.lock:
            return self.current

    def set_direction(self, direction: int) -> None:
        """
        Set the direction of the motor.
        """
        with self.lock:
            self.direction = direction

    def get_direction(self) -> int:
        """
        Get the direction of the motor.
        """
        with self.lock:
            return self.direction

    def set_position(self, position: int) -> None:
        """
        Set the position of the motor.
        """
        with self.lock:
            self.position = position

    def get_position(self) -> int:
        """
        Get the position of the motor.
        """
        with self.lock:
            return self.position

    def emergency_stop(self) -> None:
        """
        Stop the motor in case of emergency.
        """
        with self.lock:
            self.stop = True

    def monitor(self) -> None:
        """
        Monitor the motor status.
        """
        while not self.stop:
            logging.info(f"Motor status: Speed {self.get_speed()}, Temperature {self.get_temperature()}, Current {self.get_current()}, Direction {self.get_direction()}, Position {self.get_position()}")
            time.sleep(1)

    def control(self) -> None:
        """
        Control the motor using the given parameters.
        """
        try:
            while not self.stop:
                # Control logic here
                time.sleep(1)
        except Exception as e:
            logging.error(f"Motor control error: {e}")
        finally:
            self.emergency_stop()


def test_motor_control() -> None:
    """
    Unit test for the MotorControl class.
    """
    # Test code here


if __name__ == "__main__":
    motor_control = MotorControl()
    control_thread = threading.Thread(target=motor_control.control)
    monitor_thread = threading.Thread(target=motor_control.monitor)
    control_thread.start()
    monitor_thread.start()
    control_thread.join()
    monitor_thread.join()